api/requirements.txt
# langchain
# langchain_openai
langchain_core
langchain_community
docx2txt
pypdf
# faiss-cpu
python-multipart
streamlit
fastapi
uvicorn[standard]
pydantic
langchain>=0.1.0
faiss-cpu>=1.7.4 
openai>=1.12.0
langchain-openai>=0.0.5